FINERACT-1724 swagger fix

+++ integration-tests/src/test/java/org/apache/fineract/integrationtests/common/PaymentTypeDomain.java
@@ -20,13 +20,13 @@ package org.apache.fineract.integrationtests.common;
 
 public final class PaymentTypeDomain {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private Boolean isCashPayment;
     private Integer position;
 
-    public PaymentTypeDomain(final Integer id, final String name, final String description, final Boolean isCashPayment,
+    public PaymentTypeDomain(final Long id, final String name, final String description, final Boolean isCashPayment,
             final Integer position) {
         this.id = id;
         this.name = name;
@@ -36,11 +36,11 @@ public final class PaymentTypeDomain {
 
     }
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
